feat(gql): ISO GQL conformance, correctness + OOM fixes, RDF loader, decorrelation, and coverage
A consolidated set of GQL engine improvements over the prior baseline.

Conformance & features
- Broader ISO/IEC 39075 surface: CALL (imports) { subquery } scoped subqueries,
  FOR-variable binding ahead of an anchoring MATCH, external-key resolution for
  a node's .key, and postfix predicate handling (IS NULL/DIRECTED/SOURCE/DEST).

Correctness
- Resolve RETURN aliases inside postfix-IS predicates in ORDER BY (previously
  only IS LABELED was substituted; the other arms silently skipped).
- Explicit Lua-type conversion of a node property in getPropertyLua.
- Nested-EXISTS COUNT{} handling and correlated-subquery precompute.

Performance
- Decorrelate the correlated-subquery precompute.
- Cardinality-based multi-MATCH join reorder; anchor traversal on a bound node;
  bound-target semi-join for verification matches.

Robustness
- Bound unbounded shortest-path quantifiers to prevent runaway expansion (OOM).

Tooling & structure
- RDF loader.
- Modularized the GQL sources: every unit split under the file-size ceiling as
  pure code motion (value layer, parser, optimizer helpers, and their tests).
- Extensive optimizer and value-layer unit-test coverage.
- Build: pin clang to the released 22 (trunk 23 broke a pinned dependency) and
  link the test binary with lld to avoid link-time OOM.

+++ src/gql/executor/FactorNode.cpp
@@ -123,17 +123,17 @@ std::map<GqlValue, FactorNodePtr, GqlValueLess> FactorNode::partition(const std:
         for (size_t i = 0; i < children.size(); ++i) {
             auto free_vars = children[i]->freevars();
             if (free_vars.count(var)) {
-                target_idx = i;
+                target_idx = static_cast<int>(i);
                 break;
             }
         }
         if (target_idx != -1) {
-            auto target_node = children[target_idx];
+            auto target_node = children[static_cast<size_t>(target_idx)];
             auto target_parts = target_node->partition(var);
             for (auto& [val, part] : target_parts) {
                 auto product_node = std::make_shared<FactorNode>(type);
                 product_node->children = children;
-                product_node->children[target_idx] = part;
+                product_node->children[static_cast<size_t>(target_idx)] = part;
                 partition_map[val] = product_node;
             }
         }
